Extract SequencingReadType.java
